More docs & learning snippets cleanup
