+++ apps/full-stack-tests/src/hierarchies-react/learning-snipptets/Localization.test.tsx
@@ -130,9 +130,7 @@ describe("Hierarchies React", () => {
         // __PUBLISH_EXTRACT_END__
 
         const { getAllByRole } = render(<MyTreeComponent imodelAccess={access} imodelKey={imodel.key} />);
-        await waitFor(() => getAllByRole("treeitem"));
-
-        expect(getAllByRole("button", { name: "Apply hierarchy filter" })).to.not.be.empty;
+        await waitFor(() => expect(getAllByRole("button", { name: "Apply hierarchy filter" })).to.not.be.empty);
       });
 
       it("Tree renderer localization", async function () {
@@ -169,10 +167,8 @@ describe("Hierarchies React", () => {
           return <MyTreeRenderer {...state} rootNodes={rootNodes} localizedStrings={localizedStrings} />;
         }
 
-        const { getByRole, getAllByRole } = render(<MyTreeComponent imodelAccess={access} imodelKey={imodel.key} />);
-        await waitFor(() => getByRole("tree"));
-
-        expect(getAllByRole("button", { name: "Apply hierarchy filter" })).to.not.be.empty;
+        const { getAllByRole } = render(<MyTreeComponent imodelAccess={access} imodelKey={imodel.key} />);
+        await waitFor(() => expect(getAllByRole("button", { name: "Apply hierarchy filter" })).to.not.be.empty);
       });
     });
   });
