+++ packages/hierarchies/learning/imodel/PerformanceTuning.md
@@ -1,6 +1,6 @@
-# IModel hierarchy provider's performance tuning
+# iModel hierarchy provider's performance tuning
 
-While the ECSQL queries returned by [hierarchy definition](./HierarchyDefinitions.md) have the biggest impact on performance of creating a hierarchy, there are a few customization options that allow to fine-tune it based on consumers' requirements. This page lists those options.
+While the ECSQL queries returned by [hierarchy definition](./HierarchyDefinition.md) have the biggest impact on performance of creating a hierarchy, there are a few customization options that allow to fine-tune it based on consumers' requirements. This page lists those options.
 
 ## Hierarchy provider's query concurrency
